Wrote a bunch of stuff
Ok, I've been writing for a bit.

It still doesn't draw anything, but it's closer than it was.

I've got a lot more backend written, and if the front end is written we'd be fairly close to a small but working thing.

I added a global file that defines some macros in all files.

I also added a lot more stuff to everything.

Also, I had earlier decided that change instructions would be split into two instructions
	One that did the action
	And one that undid it after the duration
This didn't go very well because the calculations moved to inside the sprite.
As a result, the sprite has to keep track of stopping actions itself now.

That is all.